rename panic_decoder to address_decoder
